Fix for JRUBY-224, File.rename() should throw error when renaming to file in nonexistant directory
git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@2504 961051c9-f516-0410-bf72-c9f7e237a7b7

+++ src/org/jruby/runtime/builtin/meta/FileMetaClass.java
@@ -519,9 +519,10 @@ public class FileMetaClass extends IOMetaClass {
         oldNameString.checkSafeString();
         newNameString.checkSafeString();
         JRubyFile oldFile = JRubyFile.create(getRuntime().getCurrentDirectory(),oldNameString.toString());
-        
-        if (!oldFile.exists()) {
-        	throw getRuntime().newErrnoENOENTError("No such file: " + oldNameString);
+        JRubyFile newFile = JRubyFile.create(getRuntime().getCurrentDirectory(),newNameString.toString());
+
+        if (!oldFile.exists() || !newFile.getParentFile().exists()) {
+        	throw getRuntime().newErrnoENOENTError("No such file or directory - " + oldNameString + " or " + newNameString);
         }
         oldFile.renameTo(JRubyFile.create(getRuntime().getCurrentDirectory(),newNameString.toString()));
